// src/app/api/model/contentRatingList.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContentRating } from './contentRating';


export interface ContentRatingList { 
    results?: Array<ContentRating>;
    id?: number;
}