// src/app/api/model/keywordList.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { KeywordListItem } from './keywordListItem';


export interface KeywordList { 
    id?: number;
    keywords?: Array<KeywordListItem>;
}