// src/app/api/model/releaseDateList.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReleaseDateCountry } from './releaseDateCountry';


export interface ReleaseDateList { 
    id?: number;
    results?: Array<ReleaseDateCountry>;
}